move KeySelectionValueSelectionValueListAttribute to manager
